feat(preview): added  createImgPreview picture preview function
