+++ CHANGELOG.zh_CN.md
@@ -9,6 +9,7 @@
 - **CropperImage** `Cropper` 头像裁剪新增圆形裁剪功能
 - **CropperAvatar** 新增头像上传组件
 - **Drawer** `useDrawer`新增`closeDrawer`函数
+- **Preview** 新增`createImgPreview`图片预览函数
 
 ### 🐛 Bug Fixes
