+++ src/components/Preview/src/typing.ts
